requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.3
scikit-image==0.21.0
matplotlib==3.8.1
